(maint) Move backends into prm package
These functions are dependent on each other to the point that it makes little sense to split out
